backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.2
pdf2image==1.16.3
PyPDF2==3.0.1
Pillow==9.5.0
python-dotenv==1.0.0
pymongo==4.6.1
flask-pymongo==2.3.0
torch
torchvision
easyocr==1.7.0
gunicorn==21.2.0
